backend/requirements.txt
flask
rapidfuzz
requests
google-genai
python-dotenv
